Servicos(Admin) funcionando correctamente.

+++ demo/src/main/java/com/example/demo/Repository/ServicioRepository.java
@@ -1,12 +1,13 @@
 package com.example.demo.Repository;
 
-import com.example.demo.Model.Servicio;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 import java.util.List;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import com.example.demo.Model.Servicio;
+
 @Repository
-public interface ServicioRepository extends JpaRepository<Servicio, Long> {
-    
+public interface ServicioRepository extends JpaRepository<Servicio, Integer> {    
     List<Servicio> findByActivoTrue();
 }
